trip_ingestion_service/requirements.txt
uvicorn
fastapi
python-dotenv
aio-pika
pytest
pytest-asyncio
pymysql
sqlalchemy
cryptography
